requirements.txt
isort==5.10.1
Django==4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
pendulum==2.1.2
requests==2.26.0
psycopg2-binary==2.9.2
drf-yasg==1.20.0
PyJWT==2.3.0
coveralls==3.3.1
codacy-coverage==1.3.11
coreschema==0.0.4
coverage==6.2
whitenoise==5.3.0
celery==5.2.1
redis==4.0.2
tenacity==8.0.1
gevent==21.12.0
gunicorn==20.1.0